chatbot/requirements.txt
streamlit>=1.28.0
duckdb>=0.9.0
anthropic>=0.18.0
openai>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
plotly>=5.18.0
